requirements.txt
requests==2.31.0
pandas==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
streamlit==1.31.1